added search reset; created control container to hold shared state
