+++ src/hooks/useDataApi.js
@@ -1,25 +1,7 @@
 import { useReducer, useState, useEffect, useRef } from 'react';
 import axios from 'axios';
 import regeneratorRuntime from "regenerator-runtime";
-
-const dataFetchReducer = (state, action) => {
-  switch (action.type) {
-    case 'FETCH_INIT':
-      return { ...state, isLoading: true, isError: false };
-    case 'FETCH_SUCCESS':
-      return { 
-        ...state, 
-        isLoading: false, 
-        isError: false, 
-        numResults: action.payload.info.total,
-        results: action.payload.data,
-      };
-    case 'FETCH_FAILURE':
-      return { ...state, isLoading: false, isError: true, };
-    default:
-      throw new Error();
-  }
-};
+import dataFetchReducer from '../reducers/dataFetchReducer';
 
 const useDataApi = (initialUrl, initialData) => {
   const [url, setUrl] = useState(initialUrl);
